q_python/requirements/base.txt
fastapi>=0.95.0
uvicorn[standard]>=0.23.0
python-multipart>=0.0.6
Pillow>=10.0.0
numpy>=1.24.0
pandas>=2.0.0
pandas-ta>=0.3.14b0
opencv-python>=4.8.0
requests>=2.31.0
fredapi>=0.5.1

# News API Endpoint Dependencies (LunarCrush + FinBERT Sentiment Analysis)
# - requests: For LunarCrush API calls (already listed above)
# - python-dateutil: For date parsing in news services (LunarCrush, Stock News)
# - transformers/torch: For FinBERT sentiment analysis (already listed below)
# - python-dotenv: For loading LUNARCRUSH_API_KEY from .env file (already listed below)
python-dateutil>=2.8.2

# OCR libraries (choose one or both)
easyocr>=1.7.0
# pytesseract>=0.3.10  # Requires Tesseract OCR to be installed on system
# Face recognition libraries (choose one)
deepface>=0.0.79
# face-recognition>=1.3.0  # Requires dlib which can be difficult to install on Windows
# MediaPipe for face detection and landmarks
# NOTE: MediaPipe has limited Windows support and may not install on Windows.
# For Windows, use OpenCV (already included) or DeepFace for face detection instead.
# Uncomment below for Linux/macOS:
# mediapipe>=0.10.0

# FinGPT Model Dependencies
transformers>=4.32.0
peft>=0.5.0
torch>=2.0.0
accelerate>=0.20.0
sentencepiece>=0.1.99
huggingface-hub>=0.16.0
python-dotenv>=1.0.0
psycopg2-binary>=2.9.0